website/requirements.txt
Flask==0.10.1
Flask-Login==0.2.11
Flask-SQLAlchemy==1.0
Flask-WTF==0.9.4
Flask-Babel==0.9
Flask-Assets==0.9
Flask-Mail==0.9
celery==3.1.9
mysql-connector-python==1.1.6
boto==2.26.0
Pillow==2.3.1
cssmin
jsmin